requirements.txt
transformers==4.24.0
datasets==2.6.1
pillow==9.3.0
torch
easyocr==1.6.2
reportlab==3.6.12
gspread==5.6.2
oauth2client==4.1.3
Click==7.1.2
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pytesseract